+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Changelog
+
 ## [2.0.0] - 2024-08-08
 ### Breaking Changes
 - Changed to a new wallet factory contract address
+
+## [1.1.3] - 2024-07-24
+### New
+- Added `signUserOpWithSessionKey` function into SessionKeyValidator for sign the userOP using session key.
+- Added `sessionData` function into SessionKeyValidator to fetch session Data of particular sessionKey.
+### Breaking Changes
+- Updated `enableSessionKey`, `rotateSessionKey` and `disableSessionKey` functions for the multiple session keys per apiKey and walletAddress.
 
 ## [1.1.2] - 2024-07-10
 ### New